+++ forc-plugins/forc-doc/src/descriptor.rs
@@ -1,7 +1,10 @@
 //! Determine whether a [Declaration] is documentable.
 use crate::{
     doc::{Document, ModuleInfo},
-    render::{trim_fn_body, Context, ContextType, DocStrings, ItemBody, ItemContext, ItemHeader},
+    render::{
+        item::{components::*, context::*},
+        util::format::{code_block::trim_fn_body, docstring::DocStrings},
+    },
 };
 use anyhow::Result;
 use sway_core::{
