+++ forc-plugins/forc-doc/src/render/constant.rs
@@ -1,0 +1,3 @@
+pub(crate) const ALL_DOC_FILENAME: &str = "all.html";
+pub(crate) const INDEX_FILENAME: &str = "index.html";
+pub(crate) const IDENTITY: &str = "#";
