+++ forc-plugins/forc-doc/src/render/item/mod.rs
@@ -1,0 +1,3 @@
+mod components;
+mod context;
+mod type_anchor;
