+++ forc-plugins/forc-doc/src/render/util/mod.rs
@@ -1,0 +1,1 @@
+mod format;
